Allow running of action type tests as app user

+++ test/e2e/entity/actionTypes.spec.js
@@ -7,22 +7,8 @@ module.exports = (scope, isOperator) => {
   before(() => {
     scope = scope();
   });
-  
+
   describe('Action Types', () => {
-    it('should create an action type', async () => {
-      const payload = { name: TYPE_NAME };
-      resources.actionType = await scope.actionType().create(payload);
-
-      expect(resources.actionType).to.be.an('object');
-    });
-
-    it('should read an action type', async () => {
-      const res = await scope.actionType(resources.actionType.name).read();
-
-      expect(res).to.be.an('object');
-      expect(res.name).to.equal(resources.actionType.name);
-    });
-
     it('should read all action types', async () => {
       const res = await scope.actionType().read();
 
@@ -30,16 +16,32 @@ module.exports = (scope, isOperator) => {
       expect(res).to.have.length.gte(5);
     });
 
-    it('should update an action type', async () => {
-      const payload = { tags: ['updated'] };
-      const res = await scope.actionType(resources.actionType.name).update(payload);
-      
-      expect(res).to.be.an('object');
-      expect(res.tags).to.deep.equal(payload.tags);
-    });
+    if (isOperator) {
+      it('should create an action type', async () => {
+        const payload = { name: TYPE_NAME };
+        resources.actionType = await scope.actionType().create(payload);
 
-    it('should delete an actionType', async () => {
-      await scope.actionType(resources.actionType.name).delete();
-    });
+        expect(resources.actionType).to.be.an('object');
+      });
+
+      it('should read an action type', async () => {
+        const res = await scope.actionType(resources.actionType.name).read();
+
+        expect(res).to.be.an('object');
+        expect(res.name).to.equal(resources.actionType.name);
+      });
+
+      it('should update an action type', async () => {
+        const payload = { tags: ['updated'] };
+        const res = await scope.actionType(resources.actionType.name).update(payload);
+
+        expect(res).to.be.an('object');
+        expect(res.tags).to.deep.equal(payload.tags);
+      });
+
+      it('should delete an actionType', async () => {
+        await scope.actionType(resources.actionType.name).delete();
+      });
+    }
   });
 };
